Possibility to define custom user segments.
This commits adds possibility to customize Waffle with possibility to define
your own custom user segments.

For this you need to create a Django form, set it to WAFFLE_CUSTOM_SEGMENT
setting and implement form method called `flag_is_active`. Then in Django
admin, you can set values for fields of this form, these values will be passed
to `flag_is_active` method and this methos will decide if current user falls in
to a custom segment or not.