Use strings for cache keys
As discovered in YPlan/django-perf-rec#21, django-waffle is using `bytes` objects for cache keys. This seems to work with most of the caching infrastructure, but it's not documented as supported in the [Django cache docs](https://docs.djangoproject.com/en/1.10/topics/cache/) nor does it seem to be tested for in its test suite. Thus some cache backends may not support `bytes`, and `django-perf-rec` currently doesn't support them either.