Feature: more granular querystring flag overrides
This enables the user to choose individual flags to be overridden in
the querystring. Currently we have a binary choice (all or none) via
`WAFFLE_OVERRIDE` setting.

Adds a new `NullBooleanField` field on the `Flag` model called
`allow_override`. The default value (`None`) means that the control
depends on the global `WAFFLE_OVERRIDE` value. When set to `False`
disables the override even if `WAFFLE_OVERRIDE` is set to `True`.
Finally, when the field is set to `True` we allow querystring
overrides for this flag (as described in documentation) even if
`WAFFLE_OVERRIDE` is set to `False`.