Implements a way of branching behavior without if/else
Creates a way to avoid nesting code with if/else when changing logic
wether flag is active or not. This force the users of the lib of
creating more readable and modular code.

It also makes easier to implement automatic removal of flags in the
code base using Abstract Syntax Tree tools like astroid.