Optionally create missing Flags, Samples and Switches
Flags, Switches and Samples may be scattered throughout templates,
backend code or across mulitple frontend clients. This patch provides
a user with the option of having waffle create a missing Flag, Sample
or Switch when one is requested but does not exist.

The goal of this change is to save operators time by removing the need
to search through code for flag/switch/sample names and to reduce
communication requirements accross teams with of multiple developers
with multiple environments that might require the creation of a
flag/switch/sample in the database of every environment code is run
in.

Behavior Changes
================

Old Behavior
-----------
  1. Flag/Switch/Sample is requested
  2a. If found in db evaulate what the value should be for request
  2b. If missing return (FLAG|SWITCH|SAMPLE)_DEFAULT value

New Behavior
------------
  1. Flag/Switch/Sample is requested
  2a. If found in db evaulate what the value should be for request
  2b. Create Flag/Switch/Sample with
      value (FLAG|SWITCH|SAMPLE)_DEFAULT
  3. Return (FLAG|SWITCH|SAMPLE)_DEFAULT for this request, subsequent
  requests will be served using the value from the cache/db

Further Discussion
------------------
  - #44 (https://github.com/jsocol/django-waffle/issues/44)
  - #50 (https://github.com/jsocol/django-waffle/issues/50)

Future Work
-----------
This change intentionally narowly scoped and as such does not provide
a mechanism to set a default value other
than (FLAG|SWITCH|SAMPLE)_DEFAULT for individual objects that are not
in the database although it could be expanded in the future to do so.