Optionally create missing Flags, Samples and Switches (#301)
Flags, Switches and Samples may be scattered throughout templates,
backend code or across multiple frontend clients. This patch provides
a user with the option of having waffle create a missing Flag, Sample
or Switch when one is requested but does not exist.

The goal of this change is to save operators time by removing the need
to search through code for flag/switch/sample names and to reduce
communication requirements across teams with of multiple developers
with multiple environments that might require the creation of a
flag/switch/sample in the database of every environment code is run
in.