Flush cache after adding/removing users and groups
Currently, the cached value for a flag is not invalidated when users
or groups are added or removed. This fixes that by registering receivers
for the corresponding m2m_changed signals.

A new signals module is introduced for this purpose, and imported in the
ready method of a new AppConfig class. This is the recommended way of
defining signals in all of the currently supported versions of Django:

https://docs.djangoproject.com/en/1.11/topics/signals/