Refactor test_testutils and flush cache on delete
The functionality contained within testutils could be used with any
django test case class. This commit therefore expands the tests to cover
TestCase as well as TransactionTestCase from Django.

Doing this reveals a bug that if an object is created by the override_*
helper function then it is not flushed from the cache on deletion. This
commit adds this flush action after the delete has occurred.