Resolve flake8 errors
https://stackoverflow.com/questions/57074300/what-is-the-recommended-way-to-break-a-long-if-statement-w504-line-break-after
suggests the recommended approach to breaking long binary statements
is before the operator; we add an exclusion for that, and this
causes an eruption of line-length warnings. We suppress those too.

noqa: F401 is used to square the circle of mypy needing those
imports defined, but flake8 not seeing that they're necessary.